Configuration: added disk image create/import views
User now have more control over disk image creation and in the future can modify disk images as well.
Importing files works as well. The file will be copied to the VM bundle.
When a VM is being created, all disk images will be written to /tmp, and the directory is moved to the VM bundle during creation.
This resolves #14